Change surgical procedure terminology from SNOMED CT to NCIT (#87)
I realized that I had not revisited out terminology for the POP
procedures. Currently, we are using the SNOMED CT codes for surgical
procedures.

IMO the SNOMED CT concepts for procedure are too granular. This makes
the data input complex (too much choice or non-existing codes for
certain granularity) and makes filtering later by procedure types
impractical. A nice example is: `Lobectomy of lung with excision of part
of adjacent lobe` or `Removal of implantable venous access port using
fluoroscopic guidance`. This will results in a large spread of procedure
codes being used with not much aggregation capabilities without a
additional terminology service.
 
After some research, I propose that (as we have done with many other
data models in the new POP) we switch to the NCIT terminology. This is
much less granular, hence easier to manage. It does not have the
anatomical location integrated into the code (which is good since in POP
we collect the anatomical location in a different field).

---------

Co-authored-by: Luis Fabregas <Luis.Fabregas-Ibanez@usz.ch>